Simplify card creation and activation
